Corrected rendering of cubes.

+++ assignments/assignment5_camera/main.cpp
@@ -75,7 +75,7 @@ int main() {
 	camera.position = ew::Vec3(0.0, 0.0, 5.0);
 	camera.target = ew::Vec3(0.0, 0.0, 0.0);
 	camera.fov = 60.0;
-	camera.aspectRatio = SCREEN_WIDTH / SCREEN_HEIGHT;
+	camera.aspectRatio = (float)SCREEN_WIDTH / SCREEN_HEIGHT;
 	camera.nearPlane = 0.1;
 	camera.farPlane = 100.0;
 	camera.orthographic = true;
